removed some portions of the guest access flow in favor of the unauthed flow w/ burnettk
