+++ spiffworkflow-backend/src/spiffworkflow_backend/routes/process_api_blueprint.py
@@ -1,6 +1,8 @@
 import json
+import os
 import uuid
 from typing import Any
+from typing import TypedDict
 from uuid import UUID
 
 import flask.wrappers
@@ -21,19 +23,23 @@ from spiffworkflow_backend.background_processing.celery_tasks.process_instance_t
 )
 from spiffworkflow_backend.data_migrations.process_instance_migrator import ProcessInstanceMigrator
 from spiffworkflow_backend.exceptions.api_error import ApiError
+from spiffworkflow_backend.exceptions.error import HumanTaskAlreadyCompletedError
+from spiffworkflow_backend.exceptions.error import HumanTaskNotFoundError
+from spiffworkflow_backend.exceptions.error import UserDoesNotHaveAccessToTaskError
 from spiffworkflow_backend.exceptions.process_entity_not_found_error import ProcessEntityNotFoundError
 from spiffworkflow_backend.models.db import db
 from spiffworkflow_backend.models.human_task import HumanTaskModel
 from spiffworkflow_backend.models.human_task_user import HumanTaskUserModel
 from spiffworkflow_backend.models.principal import PrincipalModel
 from spiffworkflow_backend.models.process_instance import ProcessInstanceModel
+from spiffworkflow_backend.models.process_instance import ProcessInstanceStatus
 from spiffworkflow_backend.models.process_instance_file_data import ProcessInstanceFileDataModel
 from spiffworkflow_backend.models.process_model import ProcessModelInfo
 from spiffworkflow_backend.models.reference_cache import ReferenceCacheModel
 from spiffworkflow_backend.models.reference_cache import ReferenceSchema
-from spiffworkflow_backend.models.task import TaskModel  # noqa: F401
-from spiffworkflow_backend.services.authentication_service import AuthenticationService  # noqa: F401
+from spiffworkflow_backend.models.task import TaskModel
 from spiffworkflow_backend.services.authorization_service import AuthorizationService
+from spiffworkflow_backend.services.file_system_service import FileSystemService
 from spiffworkflow_backend.services.git_service import GitCommandError
 from spiffworkflow_backend.services.git_service import GitService
 from spiffworkflow_backend.services.jinja_service import JinjaService
@@ -42,10 +48,22 @@ from spiffworkflow_backend.services.process_instance_processor import ProcessIns
 from spiffworkflow_backend.services.process_instance_queue_service import ProcessInstanceQueueService
 from spiffworkflow_backend.services.process_instance_service import ProcessInstanceService
 from spiffworkflow_backend.services.process_model_service import ProcessModelService
+from spiffworkflow_backend.services.spec_file_service import SpecFileService
 from spiffworkflow_backend.services.task_service import TaskModelError
 from spiffworkflow_backend.services.task_service import TaskService
 
 process_api_blueprint = Blueprint("process_api", __name__)
+
+
+class TaskDataSelectOption(TypedDict):
+    value: str
+    label: str
+
+
+class ReactJsonSchemaSelectOption(TypedDict):
+    type: str
+    title: str
+    enum: list[str]
 
 
 def permissions_check(body: dict[str, dict[str, list[str]]]) -> flask.wrappers.Response:
@@ -557,3 +575,211 @@ def _get_spiff_task_from_processor(
             )
         )
     return spiff_task
+
+
+def _get_task_model_from_guid_or_raise(task_guid: str, process_instance_id: int) -> TaskModel:
+    task_model: TaskModel | None = TaskModel.query.filter_by(guid=task_guid, process_instance_id=process_instance_id).first()
+    if task_model is None:
+        raise ApiError(
+            error_code="task_not_found",
+            message=f"Cannot find a task with guid '{task_guid}' for process instance '{process_instance_id}'",
+            status_code=400,
+        )
+    return task_model
+
+
+def _get_task_model_for_request(
+    process_instance_id: int,
+    task_guid: str = "next",
+    with_form_data: bool = False,
+) -> TaskModel:
+    process_instance = _find_process_instance_by_id_or_raise(process_instance_id)
+
+    if process_instance.status == ProcessInstanceStatus.suspended.value:
+        raise ApiError(
+            error_code="error_suspended",
+            message="The process instance is suspended",
+            status_code=400,
+        )
+
+    process_model = _get_process_model(
+        process_instance.process_model_identifier,
+    )
+
+    task_model = _get_task_model_from_guid_or_raise(task_guid, process_instance_id)
+    task_definition = task_model.task_definition
+
+    can_complete = False
+    try:
+        AuthorizationService.assert_user_can_complete_task(process_instance.id, task_model.guid, g.user)
+        can_complete = True
+    except (
+        HumanTaskNotFoundError,
+        UserDoesNotHaveAccessToTaskError,
+        HumanTaskAlreadyCompletedError,
+    ):
+        can_complete = False
+
+    task_model.process_model_display_name = process_model.display_name
+    task_model.process_model_identifier = process_model.id
+    task_model.typename = task_definition.typename
+    task_model.can_complete = can_complete
+    task_model.name_for_display = TaskService.get_name_for_display(task_definition)
+    extensions = TaskService.get_extensions_from_task_model(task_model)
+
+    if with_form_data:
+        task_process_identifier = task_model.bpmn_process.bpmn_process_definition.bpmn_identifier
+        process_model_with_form = process_model
+
+        refs = SpecFileService.get_references_for_process(process_model_with_form)
+        all_processes = [i.identifier for i in refs]
+        if task_process_identifier not in all_processes:
+            top_bpmn_process = TaskService.bpmn_process_for_called_activity_or_top_level_process(task_model)
+            bpmn_file_full_path = ProcessInstanceProcessor.bpmn_file_full_path_from_bpmn_process_identifier(
+                top_bpmn_process.bpmn_process_definition.bpmn_identifier
+            )
+            relative_path = os.path.relpath(bpmn_file_full_path, start=FileSystemService.root_path())
+            process_model_relative_path = os.path.dirname(relative_path)
+            process_model_with_form = ProcessModelService.get_process_model_from_relative_path(process_model_relative_path)
+
+        form_schema_file_name = ""
+        form_ui_schema_file_name = ""
+        task_model.signal_buttons = TaskService.get_ready_signals_with_button_labels(process_instance_id, task_model.guid)
+
+        if "properties" in extensions:
+            properties = extensions["properties"]
+            if "formJsonSchemaFilename" in properties:
+                form_schema_file_name = properties["formJsonSchemaFilename"]
+            if "formUiSchemaFilename" in properties:
+                form_ui_schema_file_name = properties["formUiSchemaFilename"]
+
+        task_draft_data = TaskService.task_draft_data_from_task_model(task_model)
+
+        saved_form_data = None
+        if task_draft_data is not None:
+            saved_form_data = task_draft_data.get_saved_form_data()
+
+        task_model.data = task_model.get_data()
+        task_model.saved_form_data = saved_form_data
+        if task_definition.typename == "UserTask":
+            if not form_schema_file_name:
+                raise (
+                    ApiError(
+                        error_code="missing_form_file",
+                        message=f"Cannot find a form file for process_instance_id: {process_instance_id}, task_guid: {task_guid}",
+                        status_code=400,
+                    )
+                )
+
+            form_dict = _prepare_form_data(
+                form_file=form_schema_file_name,
+                task_model=task_model,
+                process_model=process_model_with_form,
+                revision=process_instance.bpmn_version_control_identifier,
+            )
+            _update_form_schema_with_task_data_as_needed(form_dict, task_model.data)
+            task_model.form_schema = form_dict
+
+            if form_ui_schema_file_name:
+                ui_form_contents = _prepare_form_data(
+                    form_file=form_ui_schema_file_name,
+                    task_model=task_model,
+                    process_model=process_model_with_form,
+                    revision=process_instance.bpmn_version_control_identifier,
+                )
+                task_model.form_ui_schema = ui_form_contents
+            else:
+                task_model.form_ui_schema = {}
+            _munge_form_ui_schema_based_on_hidden_fields_in_task_data(task_model.form_ui_schema, task_model.data)
+
+        # it should be safe to add instructions to the task spec here since we are never commiting it back to the db
+        extensions["instructionsForEndUser"] = JinjaService.render_instructions_for_end_user(task_model, extensions)
+
+    task_model.extensions = extensions
+    return task_model
+
+
+# originally from: https://bitcoden.com/answers/python-nested-dictionary-update-value-where-any-nested-key-matches
+def _update_form_schema_with_task_data_as_needed(in_dict: dict, task_data: dict) -> None:
+    for k, value in in_dict.items():
+        if "anyOf" == k:
+            # value will look like the array on the right of "anyOf": ["options_from_task_data_var:awesome_options"]
+            if isinstance(value, list):
+                if len(value) == 1:
+                    first_element_in_value_list = value[0]
+                    if isinstance(first_element_in_value_list, str):
+                        if first_element_in_value_list.startswith("options_from_task_data_var:"):
+                            task_data_var = first_element_in_value_list.replace("options_from_task_data_var:", "")
+
+                            if task_data_var not in task_data:
+                                message = (
+                                    "Error building form. Attempting to create a selection list with options from"
+                                    f" variable '{task_data_var}' but it doesn't exist in the Task Data."
+                                )
+                                raise ApiError(
+                                    error_code="missing_task_data_var",
+                                    message=message,
+                                    status_code=500,
+                                )
+
+                            select_options_from_task_data = task_data.get(task_data_var)
+                            if select_options_from_task_data == []:
+                                raise ApiError(
+                                    error_code="invalid_form_data",
+                                    message=(
+                                        "This form depends on variables, but at least one variable was empty. The"
+                                        f" variable '{task_data_var}' must be a list with at least one element."
+                                    ),
+                                    status_code=500,
+                                )
+                            if isinstance(select_options_from_task_data, str):
+                                raise ApiError(
+                                    error_code="invalid_form_data",
+                                    message=(
+                                        "This form depends on enum variables, but at least one variable was a string."
+                                        f" The variable '{task_data_var}' must be a list with at least one element."
+                                    ),
+                                    status_code=500,
+                                )
+                            if isinstance(select_options_from_task_data, list):
+                                if all("value" in d and "label" in d for d in select_options_from_task_data):
+
+                                    def map_function(
+                                        task_data_select_option: TaskDataSelectOption,
+                                    ) -> ReactJsonSchemaSelectOption:
+                                        return {
+                                            "type": "string",
+                                            "enum": [task_data_select_option["value"]],
+                                            "title": task_data_select_option["label"],
+                                        }
+
+                                    options_for_react_json_schema_form = list(
+                                        map(
+                                            map_function,
+                                            select_options_from_task_data,
+                                        )
+                                    )
+
+                                    in_dict[k] = options_for_react_json_schema_form
+        elif isinstance(value, dict):
+            _update_form_schema_with_task_data_as_needed(value, task_data)
+        elif isinstance(value, list):
+            for o in value:
+                if isinstance(o, dict):
+                    _update_form_schema_with_task_data_as_needed(o, task_data)
+
+
+def _munge_form_ui_schema_based_on_hidden_fields_in_task_data(form_ui_schema: dict | None, task_data: dict) -> None:
+    if form_ui_schema is None:
+        return
+    if task_data and "form_ui_hidden_fields" in task_data:
+        hidden_fields = task_data["form_ui_hidden_fields"]
+        for hidden_field in hidden_fields:
+            hidden_field_parts = hidden_field.split(".")
+            relevant_depth_of_ui_schema = form_ui_schema
+            for ii, hidden_field_part in enumerate(hidden_field_parts):
+                if hidden_field_part not in relevant_depth_of_ui_schema:
+                    relevant_depth_of_ui_schema[hidden_field_part] = {}
+                relevant_depth_of_ui_schema = relevant_depth_of_ui_schema[hidden_field_part]
+                if len(hidden_field_parts) == ii + 1:
+                    relevant_depth_of_ui_schema["ui:widget"] = "hidden"
