+++ spiffworkflow-backend/tests/spiffworkflow_backend/integration/test_public_controller.py
@@ -1,9 +1,11 @@
 import json
+import re
 
 from flask import Flask
 from flask.testing import FlaskClient
 from spiffworkflow_backend.models.process_instance import ProcessInstanceModel
 from spiffworkflow_backend.models.process_instance import ProcessInstanceStatus
+from spiffworkflow_backend.models.task import TaskModel  # noqa: F401
 from spiffworkflow_backend.services.authorization_service import AuthorizationService
 from spiffworkflow_backend.services.authorization_service import GroupPermissionsDict
 
@@ -161,3 +163,102 @@ class TestPublicController(BaseTest):
 
         process_instance = ProcessInstanceModel.query.filter_by(id=process_instance_id).first()
         assert process_instance.status == ProcessInstanceStatus.user_input_required.value
+
+    def test_can_complete_complete_a_guest_task(
+        self,
+        app: Flask,
+        client: FlaskClient,
+        with_db_and_bpmn_file_cleanup: None,
+    ) -> None:
+        admin_user = self.find_or_create_user("admin")
+        group_info: list[GroupPermissionsDict] = [
+            {
+                "users": [],
+                "name": app.config["SPIFFWORKFLOW_BACKEND_DEFAULT_PUBLIC_USER_GROUP"],
+                "permissions": [{"actions": ["create", "read", "update"], "uri": "/public/*"}],
+            },
+            {
+                "users": [admin_user.username],
+                "name": "admin",
+                "permissions": [{"actions": ["all"], "uri": "/*"}],
+            },
+        ]
+        AuthorizationService.refresh_permissions(group_info)
+
+        process_group_id = "my_process_group"
+        process_model_id = "test-allow-guest"
+        bpmn_file_location = "test-allow-guest"
+        process_model = self.create_group_and_model_with_bpmn(
+            client,
+            admin_user,
+            process_group_id=process_group_id,
+            process_model_id=process_model_id,
+            bpmn_file_location=bpmn_file_location,
+        )
+
+        admin_headers = self.logged_in_headers(admin_user)
+        response = self.create_process_instance_from_process_model_id_with_api(client, process_model.id, admin_headers)
+        assert response.json is not None
+        process_instance_id = response.json["id"]
+
+        response = client.post(
+            f"/v1.0/process-instances/{self.modify_process_identifier_for_path_param(process_model.id)}/{process_instance_id}/run",
+            headers=admin_headers,
+        )
+        assert response.status_code == 200
+
+        task_model = TaskModel.query.filter_by(process_instance_id=process_instance_id, state="READY").first()
+        assert task_model is not None
+        first_task_guid = task_model.guid
+
+        response = client.get(
+            f"/v1.0/public/tasks/{process_instance_id}/{first_task_guid}",
+        )
+        assert response.status_code == 200
+        assert response.json is not None
+        assert response.json["form"] == {"form_schema": None, "form_ui_schema": None, "instructions_for_end_user": ""}
+        assert response.json["confirmation_message_markdown"] is None
+        assert response.json["task_guid"] == first_task_guid
+        assert response.json["process_instance_id"] == process_instance_id
+
+        headers_dict = dict(response.headers)
+        assert "Set-Cookie" in headers_dict
+        cookie = headers_dict["Set-Cookie"]
+        cookie_split = cookie.split(";")
+        access_token = [cookie for cookie in cookie_split if cookie.startswith("access_token=")][0]
+        assert access_token is not None
+        re_result = re.match(r"^access_token=[\w_\.-]+$", access_token)
+        assert re_result is not None
+        user_header = {"Authorization": "Bearer " + access_token.split("=")[1]}
+
+        response = client.put(
+            f"/v1.0/public/tasks/{process_instance_id}/{first_task_guid}?execution_mode=synchronous",
+            data="{}",
+            content_type="application/json",
+            headers=user_header,
+        )
+        assert response.status_code == 200
+        assert response.json is not None
+        assert response.json["form"] == {"instructions_for_end_user": "We have instructions."}
+        assert response.json["confirmation_message_markdown"] is None
+        assert response.json["task_guid"] is not None
+        assert response.json["task_guid"] != first_task_guid
+        assert response.json["process_instance_id"] == process_instance_id
+
+        second_task_guid = response.json["task_guid"]
+        response = client.put(
+            f"/v1.0/public/tasks/{process_instance_id}/{second_task_guid}?execution_mode=synchronous",
+            data="{}",
+            content_type="application/json",
+            headers=user_header,
+        )
+        assert response.status_code == 200
+        assert response.json is not None
+        assert response.json["form"] is None
+        assert response.json["confirmation_message_markdown"] == "You have completed the task."
+        assert response.json["task_guid"] is None
+        assert response.json["process_instance_id"] == process_instance_id
+
+        process_instance = ProcessInstanceModel.query.filter_by(id=process_instance_id).first()
+        assert process_instance is not None
+        assert process_instance.status == ProcessInstanceStatus.complete.value
